changes made list in description
- create seprate layouts
- title added
- login credentials added default
- all interfaces move to single file in utils
- todoz text updated
- textual changes
- theme provider move to StoreProvider
- extra code removed
- comments added on each file
