+++ src/apis/auth/authApis.ts
@@ -1,10 +1,14 @@
-import { SignInFormValueState, SignUpFormValueState } from "@/app/auth/config"
+// I have used AXIOS for api call 
+// we can do this using thunk or saga but i think this is not as much complex to go with saga or thunk 
+// and we can also use RTK async thunk but because Axios is popular and most used one that's why i used it
+
+
+import { SignInFormValueState, SignUpFormValueState } from '@/utils/constants/interfaces';
 import axios from "axios"
 import { SIGN_IN_API_URL, SIGN_UP_API_URL } from "../apiConstants"
 import { errorHandler } from "@/utils/helpers/apis"
 
 export const signInApi = async (payload: SignInFormValueState) => {
-    // chainging email to username because https://dummyjson.com/docs/auth is using username
     const payloadWithUserName = {
         username: payload?.email,
         password: payload?.password,
