--- src/app/auth/config.ts
@@ -1,18 +1,0 @@
-export interface SignInFormValueState {
-    email: string;
-    password: string | number;
-}
-
-export interface SignUpFormValueState {
-
-    // email: string;
-    // password: string | number;
-    // confirmPassword: string | number;
-    // age: number | string;
-    
-    // commenting above because on dummyjson api endpoint required below payload to create a user     
-
-    firstName: string;
-    lastName: string;
-    age: string;
-}
