+++ src/components/TodoComponents/CreateTodo/CreateTodo.tsx
@@ -8,12 +8,7 @@ import { addTodoApi } from "@/apis/todos/todoApis"
 import { addTodo } from "@/store/slices/todoSlice"
 import { toaster } from "@/utils/helpers/toaster"
 import { NEW_TODO_ADDED } from "@/utils/constants/messages"
-
-export interface AddTodoPayload {
-    todo: string,
-    completed: boolean;
-    userId: number | string;
-}
+import { AddTodoPayload } from "@/utils/constants/interfaces"
 
 const CreateTodo = () => {
