+++ src/store/slices/selectors/todo.selector.ts
@@ -1,3 +1,5 @@
+// i have declared these selectors here so i can directly use them
+
 import { RootState } from "../../store";
 export const todosList = (state: RootState) => state.todo.todos;
 export const totalTodos = (state: RootState) => state.todo.total;
