+++ src/utils/helpers/apis.ts
@@ -1,3 +1,5 @@
+// just created this files to use helper methods 
+
 import { isAxiosError } from "axios";
 import { ERR_OCCUR } from "../constants/messages";
