+++ src/utils/routes-path.ts
@@ -1,3 +1,5 @@
+// this is my route file
+
 export const URL_HOME = "/";
 export const URL_SIGN_IN = "/auth/sign-in";
 export const URL_SIGN_UP = "/auth/sign-up";
